fix разделение функций
